style(common): move config classes
